tests/submission/W3C/domhtml/HTMLAreaElement08.html

<!DOCTYPE html>
<html><head>
<meta charset="utf-8">
<title>NIST DOM HTML Test - Area</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../common/domtests-common.js"></script>
<link href="http://www.w3.org/TR/html5/the-area-element.html#dom-area-target" rel="help">
<link rel="author" title="NIST">
</head>
<body>
<p>
<map id="mapid" name="mapid">
<area alt="Domain" href="./files/dletter.html" target="dynamic"></area>
</map>
</p>


<script>
var domtest = async_test("The target specifies the frame to render the resource in.");

on_event(window, "load", domtest.step_func(function (dom_event) {
    var nodeList;
      var testNode;
      var vtarget;
      var doc;
      doc = window.document;
      nodeList = doc.getElementsByTagName("area");
      assert_equals(nodeList.length,1);
testNode = nodeList.item(0);
      vtarget = testNode.target;

      assert_equals(vtarget,"dynamic");
       
    domtest.done();
}));

</script>
<p>The target specifies the frame to render the resource in.   

    Retrieve the target attribute and examine it's value.</p>
<div id="log"></div>
</body></html>
